requirements.txt
requests
python-dotenv
datetime
inquirer
rich